API Expansions and Removals | Extracted editor logic from Flowchart
Expansions:
- Command has an overridable SkipExecution bool (returned Enabled by default) to make it clear when it's meant to serve a purpose other than executing logic
  - Label and Comment override it to always return true, for obvious reasons
- Added IVariableSource interface requiring a method for resetting the vars
- Added ICommandResetter interface
- Renames:
  - OnBlockStart -> BlockExecStart

Removals:
- BlockReference no longer holds a reference to a Block. Refactored the logic in it accordingly.
  - Much like VariableReference, it doesn't need to do its job through hard refs; storing config for
  fetching is what it's supposed to do
- Removed ColorCommands editor option
  - Why would you _not_ want commands colored?

Editor logic extracted from Flowchart:
- See the FlowchartEditorQol class
- FlowchartGlobalDefaults is the renamed FlowchartDefaultConfig.
  - The new name communicates the intent better.

Other:
- Renamed Block's command list to _legacyCommandList since eventually, we're
going to revamp Commands to be pocos
- Block and Command Ids are now bytes instead of ushorts

+++ Assets/Hyphlow/Runtime/Commands/ICommandSource.cs
@@ -14,7 +14,7 @@ namespace AtMycelia.Hyphlow
         /// Returns true if the Command was successfully removed, false if one
         /// with the given id was not found in this source.
         /// </summary>
-        bool RemoveCommandWithId(ushort id, bool triggerSignals);
+        bool RemoveCommandWithId(byte id, bool triggerSignals);
 
         /// <summary>
         /// Removes all commands from this source. Returns true if any Commands were removed,
@@ -27,7 +27,7 @@ namespace AtMycelia.Hyphlow
     public interface IHasCommands
     {
         IReadOnlyList<ICommand> Commands { get; }
-        ICommand GetCommandWithId(ushort id);
+        ICommand GetCommandWithId(byte id);
         bool Contains(ICommand cmd);
     }
